requirements.txt
aenum==3.1.15
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
dataclasses-json==0.6.7
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
marshmallow==3.26.1
mypy-extensions==1.0.0
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
sentinelhub==3.11.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
tifffile==2025.1.10
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
utm==0.8.0
